requirements.txt
google-genai
streamlit
numpy
faiss-cpu
google-api-core
google-generativeai